CUDA compilation update. Updated MUNDY_THROW_ASSERT to have separate host/device functionality. Added fmt to overcome the inability to use pipe in combination with these new macros. Made MUNDY_THROW_ASSERT debug only and MUNDY_THROW_REQUIRE as an always check. Updated MundyMath to use explicit templates rather than auto template type due to CUDA compatability. Fixed an infinite recursion in MundyMeta due to print statements calling sprint statements.

+++ mundy/mesh/include/mundy_mesh/utils/SelectorEval.hpp
@@ -100,11 +100,11 @@ class SelectorEval {
   }
 
   int get_node_count() const {
-    return node_ptrs_.size();
+    return static_cast<int>(node_ptrs_.size());
   }
 
   int get_result_buffer_size() const {
-    return result_buffer_.size();
+    return static_cast<int>(result_buffer_.size());
   }
 
   int get_head_node_index() const;
